Adapt NetBSD code to signature change.

+++ netbsd/cpu.cc
@@ -26,12 +26,12 @@
 #include "getsysctl.h"
 #include "cpu.h"
 
-uint8_t get_cpu_count()
+uint32_t get_cpu_count()
 {
   int cpu_count = 0;
   GETSYSCTL( "hw.ncpu", cpu_count );
 
-  return static_cast<uint8_t>( cpu_count );
+  return static_cast<uint32_t>( cpu_count );
 }
 
 float cpu_percentage( unsigned int cpu_usage_delay )
